Allow to suppress removal of symbolic links to unneeded external plugins
When building, symlinks got removed/readded underneath
"gerrit/plugins" as needed. This confused Eclipse's automatic updates,
if it watched the "gerrit" directory. By allowing to suppress the
removal of symbolic links, the project layout does not change so
often, and Eclipse no longer gets confused. But on the other hand, the
plugins are no longer that well isolated during building.

Hence, the default behaviour is left unchanged and the plugins are as
well isolated as possible. But for devs, the build behaviour can be
made more Eclipse friendly.

Change-Id: Iaa98a9aec3574290ae2afb7aceb9ea80c094604e