Avoiding considering extra plugins as bundled if they are linked in
When not removing plugin links, it may be that a separate plugin is
linked into the gerrit plugin directory at build start. Since when
building bundled plugins we only looked for directories (and did not
check whether or not those directories are symlinks), those separate
plugins got erroneous picked up already when building bundled
plugins. And they've been run again when building the separate
plugins.

To avoid treating already linked in separate plugins as bundled
plugins, and to stop building such plugins twice, we now check for
symbolic links when trying to find bundled plugins.

Change-Id: I3ae150c4d470378ac02059326f7dc753ae25efa4