If build dies after having written all artifacts, report it as died nonetheless
If the build died after having written all artifacts, the build's
status itself was (as expected) "died". But the overview page reported
it as "ok", as that was the infered status from the artifact group
status text file. We now report such builds as failed also in the
overview index pages.

Change-Id: I11c59995bb6178f4e748979a3fefba9208bb5f17