Fix depth when finding files for which to compute checksum
Combining *, maxdepth 1 and -type f not only considered files in the
current directory, but also files directly in directories in the
current direcotry (* resolves the directory in the current directory,
and -maxdepth 1 on top of that yields the file in the directory of the
current directory). We cannot trade * for ., as then we'd get
filenames like './index.html' (leading './') in the sha1sums.txt,
which we do not want. Hence, we adapt the maxdepth.

Change-Id: Ic0ce368d38ec97e4c8e2608ddb72ad1d1b3b4783