In overview html's group columns, link to first non-ok artifact
Up to now, the link went to the artifact group regardless of whether
the group's artifacts were ok or not. Since the first failed/broken
artifact is exposed per group (see the parent commit), we use this
information to directly link to the first failed/broken artifact per
group. If a group is fully ok, we keep the link to the group itself.

Change-Id: Id393b9ed18e1440f8a57aea078207a163c8ece37