Swap sort order of 'latest' and 'latest-ok' in overview htmls
The first (resp. last) row of the build overview table was
latest-ok. This means that the first (resp. last) row was always
green, regardless of whether or not the latest build was ok. This sort
order does not expose breakage of the latest build well enough. Hence,
we swap the order of 'latest' and 'latest-ok'. Thereby, if the latest
build is not ok, the first (resp. last) row in the table is no longer
green, and the breakage is better visible.

Change-Id: Ie204d0786d1091aa1afb2434a9f93e7a42346817